\name{litc_package-package}
\alias{litc_package-package}
\alias{litc_package}
\docType{package}
\title{
LITC package
}
\description{
More about what it does (maybe more than one line)
}
\details{
\tabular{ll}{
Package: \tab litc_package\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-08-06\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}

}
\author{
aaabbb
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
